fix: remove unnecessary types
