+++ packages/offix-client/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./ApolloOfflineClient";
-export * from "./apollo/ApolloOfflineClient";
+export * from "./apollo/ApolloOfflineTypes";
 export * from "./apollo/conflicts/ConflictLink";
 export * from "offix-offline";
 export * from "offix-conflicts-client";
